requirements.txt
ipykernel==6.24.0
jupyter==1.0.0
notebook==6.5.2
pandas==2.0.3
numpy==1.25.1
matplotlib==3.7.2
requests==2.28.2
python-dotenv==1.0.0
web3==6.0.0
gdown==5.2.0

ipython==8.14.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.3.1
jupyter_server==2.1.0
jupyterlab==3.5.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
pytz==2022.7.1
six==1.16.0